메인게시판

+++ board/src/main/java/com/study/board/repository/BoardRepository.java
@@ -1,9 +1,9 @@
 package com.study.board.repository;
 
-import com.study.board.entity.Board;
+import com.study.board.model.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BoardRepository extends JpaRepository<Board, Integer> {
+public interface BoardRepository extends JpaRepository<Board, Long> {
 }
